requirements.txt
streamlit==1.37.1
pandas==2.2.2
plotly-express==0.4.1